backend/requirements.txt
amqp==5.2.0
anyio==4.4.0
APScheduler==3.10.4
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
django-filter==24.2
django-modeltranslation==0.19.5
django-simple-history==3.7.0
django-sql-utils==0.7.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
flower==2.0.1
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
idna==3.7
kombu==5.3.7
nest-asyncio==1.6.0
openpyxl==3.1.4
outcome==1.3.0.post0
pillow==10.3.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.8.0
PySocks==1.7.1
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.3
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
tornado==6.4.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
vine==5.1.0
watchdog==4.0.1
wcwidth==0.2.13
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4.post2
